Revert "Rename kernel_iface to kernel_api."
This reverts commit 47a5135ef28679dae3c730ddf1d7a40c20d2cedf.
