+++ numba_dpex/kernel_iface/__init__.py
@@ -3,10 +3,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 """
-The kernel_api module provides a set of Python classes and functions that are
-analogous to the C++ SYCL API. The kernel_api module is meant to allow
+The kernel_iface provides a set of Python classes and functions that are
+analogous to the C++ SYCL API. The kernel_iface API is meant to allow
 prototyping SYCL-like kernels in pure Python before compiling them using
-numba_dpex.
+numba_dpex.kernel.
 """
 
 from .atomic_ref import AtomicRef
