+++ numba_dpex/kernel_iface/memory_enums.py
